improve roles seed
